requirements.txt
litellm
pyyaml
python-dotenv
termcolor
